fastapi_app/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-jose[cryptography]>=3.3
httpx>=0.27
pydantic>=2.6
